// src/app/components/shared/models/order.ts
export class Order {
  id: number;
  processedOrder: boolean;
  allowShipping: boolean;
  sendAllProduct: boolean;
  orderNumber?: string;
  idSeller: number;
  nameSeller?: string;
  nitSeller?: string;
  idChannel: number;
  channel?: string;
  dateOrder: string;
  idStatusOrder: number;
  statusOrder?: string;
  costTotalOrder: number;
  costTotalShipping: number;
  commission: number;
  dateMaxDeliveryOrder: string;
  typeDespatchOrder?: string;
  identificationCard?: string;
  clientName?: string;
  clientTelephone?: string;
  clientAddress?: string;
  products?: (ProductsEntity)[] | Array<ProductsEntity>;
}

/**
 * Información del cliente
 * @export
 * @class InfoClient
 */
export class InfoClient {
  identificationCard?: string;
  clientName?: string;
  clientTelephone?: string;
  clientAddress?: string;
}

/**
 * Información del producto
 * @export
 * @class ProductsEntity
 */
export class ProductsEntity {
  idDetailProduct: number;
  idPicking: number;
  id: number;
  nameProduct?: string;
  productId?: string;
  sku?: number;
  ean?: string;
  checkProductToSend: boolean;
  reference?: string;
  idStatusProduct: number;
  statusProduct: string;
  quantity: number;
  costProduct: number;
  costShippingProduct: number;
  commission: number;
  dateMaxDeliveryProduct: string;
  typeDespatchProduct?: string;
  tracking?: string;
  carrier?: string;
  dateDelivery: string;
  enviosExito: boolean;
  fulfillment: boolean;
}
// ...
export class Billing {
  id: number;
  idBillingPicking: number;
  idSeller: number;
  billingNumber: string;
  orderNumber: string;
  payOrderNumber: string;
  concept: string;
  paymentDate: string;
  commission: number;
  billingTotal: number;
  detail?: (DetailEntity)[] | Array<DetailEntity>;
  paginationToken: string;
}


/**
 * Include in Billing
 * @export
 * @class DetailEntity
 */
export class DetailEntity {
  id: number;
  idBillingDetailPicking: number;
  detailName: string;
  ean: string;
  quantity: number;
  commission: number;
  price: number;
}

/**
 * Estructura empleada para el proceso luego de finalizar la carga de las guías
 * @export
 * @class FinishUploadInformation
 */
export class FinishUploadInformation {
  totalProcess: number;
  error: number;
  successful: number;
  listError?: (DetailFinishUploadInformation)[] | Array<DetailFinishUploadInformation>;
}

/**
 * DetailFinishUploadInformation
 * @export
 * @class DetailFinishUploadInformation
 */
export class DetailFinishUploadInformation {
  orderNumber: string;
  sku: string;
  message: string;
}

/**
 * Estructura para las ordenes pendientes
 * @export
 * @class PendingDevolutionsModel
 */
export class OrderDevolutionsModel {
  id: string;
  idSeller: number;
  reversionRequestId: string;
  orderNumber: string;
  reversionRequestStatusId: number;
  reversionRequestStatus: string;
  reversionRequestTypeId: number;
  reversionRequestType: string;
  reversionRequestReasonId: number;
  clientObservation: string;
  reversionRequestReason: string;
  creationDate: string;
  maximumDeliveryDate: string;
  orderDate: string;
  typeReturnId: number;
  typeReturn: string;
  clientName: string;
  identificationCard: string;
  clientAddress: string;
  clientTelephone: string;
  sellerObservationReversionRequestRefuse?: null;
  sacObservationReversionRequestRefuse?: null;
  sellerObservationReceiptRefuse?: null;
  sacObservationReceiptRefuse?: null;
  reversionRequestDetailViewModel: ReversionRequestDetailViewModel;
}

/**
 * ReversionRequestDetailViewModel
 * @export
 * @class ReversionRequestDetailViewModel
 */
export class ReversionRequestDetailViewModel {
  reversionRequestDetailId: string;
  detailSubOrderMarketplaceId: string;
  productName: string;
  sku: string;
  reference: string;
  requestedAmount: number;
  reverseAmount: number;
}